+++ test/libsolidity/syntaxTests/types/address/address_to_contract_receive.sol
@@ -1,6 +1,6 @@
 contract C {
   function f() public pure returns (C c) {
-    c = C(address(uint160(2)));
+    c = C(address(2));
   }
   receive() external payable {
   }
